backend/requirement.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
Flask-Limiter==3.5.0
python-socketio==5.11.0
python-engineio==4.7.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Werkzeug==3.0.1
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
onnxruntime==1.16.3
insightface==0.7.3
reportlab==4.0.7
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3